+++ src/main/java/org/marble/level/LevelStatement.java
@@ -1,6 +1,6 @@
 package org.marble.level;
 
-import javax.vecmath.Vector3d;
+import com.jme3.math.Vector3f;
 
 import com.google.common.base.Objects;
 import com.google.common.base.Optional;
@@ -210,11 +210,11 @@ public abstract class LevelStatement {
      */
     static class Position extends LevelStatement {
         private final String name;
-        private final Vector3d position;
+        private final Vector3f position;
         private final Optional<String> relativeTo;
 
         public Position(final Integer location, final String name,
-                final Vector3d position, final Optional<String> relativeTo) {
+                final Vector3f position, final Optional<String> relativeTo) {
             super(location);
             this.name = name;
             this.position = position;
@@ -242,7 +242,7 @@ public abstract class LevelStatement {
         /**
          * The position to use when positioning the entity.
          */
-        public Vector3d getPosition() {
+        public Vector3f getPosition() {
             return position;
         }
